unify metrics creation
